requirements.txt
bottle==0.12.25
gunicorn==22.0.0
libsql-experimental==0.0.34
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
